RT work on normilising the application state, code cleanup and refactoring
