+++ src/client/package.json
@@ -87,6 +87,7 @@
     "redux-observable": "0.17.0",
     "redux-persist": "5.6.5",
     "redux-persist-crosstab": "3.7.0",
+    "reselect": "3.0.1",
     "rxjs": "5.4.2",
     "ua-parser-js": "0.7.10",
     "ws": "0.8.1"
