+++ src/client/src/configureStore.ts
@@ -3,15 +3,15 @@ import { createEpicMiddleware, combineEpics } from 'redux-observable'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import { persistStore } from 'redux-persist'
 
-import rootReducer from './combineReducers'
+import rootReducer from './redux/reducers/combineReducers'
 import { blotterServiceEpic } from './ui/blotter/blotterOperations'
-import { referenceServiceEpic } from './referenceOperations'
-import { pricingServiceEpic } from './pricingOperations'
+import { referenceServiceEpic } from './redux/actions/referenceDataActions'
+import { pricingServiceEpic } from './redux/actions/pricingActions'
 import { analyticsServiceEpic } from './ui/analytics/analyticsOperations'
-import { compositeStatusServiceEpic } from './compositeStatusServiceOperations'
+import { compositeStatusServiceEpic } from './redux/reducers/compositeStatusServiceOperations'
 import { connectionStatusEpicsCreator } from './connectionStatusOperations'
 import { popoutEpic } from './popoutEpic'
-import { spotTileEpicsCreator } from './ui/spotTile/spotTileOperations'
+import { spotTileEpicsCreator } from './redux/actions/spotTileActions'
 import { footerEpic } from './ui/footer/FooterOperations'
 
 const epicMiddleware = (referenceDataService, blotterService, pricingService, analyticsService, compositeStatusService, executionService, openFin) => createEpicMiddleware(
