+++ src/client/src/redux/reducers/compositeStatusServiceOperations.ts
@@ -1,7 +1,7 @@
 import { createAction, handleActions } from 'redux-actions'
 import * as _ from 'lodash'
-import { ACTION_TYPES as REF_ACTION_TYPES } from './referenceOperations'
-import { ServiceStatus, ServiceInstanceStatus } from './types/'
+import { ACTION_TYPES as REF_ACTION_TYPES } from '../actions/referenceDataActions'
+import { ServiceStatus, ServiceInstanceStatus } from '../../types/index'
 
 export enum ACTION_TYPES {
   COMPOSITE_STATUS_SERVICE = '@ReactiveTraderCloud/COMPOSITE_STATUS_SERVICE',
