+++ src/client/src/redux/reducers/spotTileDataReducer.ts
@@ -1,0 +1,40 @@
+import { ACTION_TYPES } from '../actions/spotTileActions'
+import { buildNotification } from '../../ui/notification/notificationUtils'
+
+const updateSpotTile = (state, symbol, value) => {
+  return {
+    ...state,
+    [symbol]: {
+      ...state[symbol],
+      ...value,
+    },
+  }
+}
+
+export const spotTileDataReducer = (state: any = {}, action) => {
+
+  const { type, payload } = action
+  switch (type) {
+    case ACTION_TYPES.DISPLAY_CURRENCY_CHART:
+      return updateSpotTile(state, payload.symbol, { currencyChartIsOpening: true })
+
+    case ACTION_TYPES.CURRENCY_CHART_OPENED:
+      return updateSpotTile(state, payload, { currencyChartIsOpening: false })
+
+    case ACTION_TYPES.EXECUTE_TRADE:
+      return updateSpotTile(state, payload.CurrencyPair, { isTradeExecutionInFlight: true })
+
+    case ACTION_TYPES.TRADE_EXECUTED:
+      const symbol = payload.hasError ? payload.trade.CurrencyPair : payload.trade.currencyPair.symbol
+      return updateSpotTile(state, symbol, {
+        hasError: payload.hasError ,
+        isTradeExecutionInFlight: false,
+        notification: buildNotification(payload.trade, payload.error)})
+
+    case ACTION_TYPES.DISMISS_NOTIFICATION:
+      return updateSpotTile(state, payload.symbol, { notification: null })
+
+    default:
+      return state
+  }
+}
