+++ src/client/src/ui/spotTile/__tests__/spotTile.spec.jsx
@@ -31,7 +31,6 @@ const spotTileProps = {
   // },
   notional: 500,
   priceStale: false,
-  pricingConnected: true,
   title: 'GBP / USD',
 }
