+++ src/client/src/ui/spotTile/priceButton/PriceButton.tsx
@@ -3,17 +3,15 @@ import * as classnames from 'classnames'
 
 import { Direction } from '../../../types'
 import './PriceButtonStyles.scss'
+import { CurrencyPair } from "../../../types/currencyPair";
+import { Rate } from "../../../types/rate";
 
 interface PriceButtonProps {
   className: string
   direction: Direction
-  rate: {
-    pips: any
-    bigFigure: number
-    pipFraction: number
-    rawRate: number,
-  },
+  rate: Rate
   onExecute: () => void
+  currencyPair: CurrencyPair
 }
 
 const renderPips = (pips: number) => pips.toString().length === 1 ? `0${pips}` : pips
