+++ src/client/src/ui/utils/mapToPropsSelectorFactory.ts
@@ -1,0 +1,7 @@
+import { createSelectorCreator, defaultMemoize } from 'reselect'
+import * as _ from 'lodash'
+
+export const createDeepEqualSelector = createSelectorCreator<any>(
+  defaultMemoize,
+  _.isEqual
+)
