+++ src/client/src/ui/workspace/WorkspaceContainer.tsx
@@ -1,17 +1,22 @@
 import * as _ from 'lodash'
 import * as React from 'react'
-
 import './WorkspaceContainerStyles.scss'
 import { connect } from 'react-redux'
 import RegionWrapper from '../../regions/RegionWrapper'
 import ConnectedSpotTileContainer from '../spotTile/SpotTileContainer'
+import { createDeepEqualSelector } from '../utils/mapToPropsSelectorFactory'
+
+const getSpotTileKeys = createDeepEqualSelector(
+  (state:any) => Object.keys(state.currencyPairs),
+  (spotTilesKeys) => spotTilesKeys
+)
 
 interface WorkspaceContainerOwnProps {
 
 }
 
 interface WorkspaceContainerStateProps {
-  spotTiles: any
+  spotTileKeys: string[]
 }
 
 interface WorkspaceContainerDispatchProps {
@@ -26,22 +31,22 @@ type WorkspaceContainerProps =
 export class WorkspaceContainer extends React.Component<WorkspaceContainerProps, {}> {
 
   render() {
-    const items = this.props.spotTiles && _.values(this.props.spotTiles)
-
     return <div className="shell__workspace">
       <div className="workspace-region">
-        {this.renderItems(items)}
+        {this.renderItems()}
       </div>
     </div>
   }
 
-  renderItems(items) {
+  renderItems() {
 
-    if (!items.length) {
+    const { spotTileKeys } = this.props
+
+    if (!spotTileKeys || spotTileKeys.length === 0) {
       return <div className="workspace-region__icon--loading"><i className="fa fa-5x fa-cog fa-spin"/></div>
     }
 
-    return Object.keys(this.props.spotTiles)
+    return spotTileKeys
       .map(key => (
         <RegionWrapper key={key} region={key}>
           <div className="workspace-region__item">
@@ -52,8 +57,10 @@ export class WorkspaceContainer extends React.Component<WorkspaceContainerProps,
   }
 }
 
-function mapStateToProps({ spotTiles }) {
-  return { spotTiles }
+function mapStateToProps(state: any) {
+  return {
+    spotTileKeys: getSpotTileKeys(state)
+  }
 }
 
 export default connect(mapStateToProps)(WorkspaceContainer)
